requirements.txt
google-cloud-storage==2.10.0
google-auth==2.23.4